refactor(settings): move judge accessors, consts, validators to settings_judge.go (PRD #1021 M3)
